improve UI and fix , don't ask cam/mic permissions for play mode
