+++ example/index.js
@@ -1,8 +1,5 @@
 import { AppRegistry } from 'react-native';
-import App from './src/App';
-//import App from './src/Play';
-//import App from './src/Peer';
-//import App from './src/Conference';
 import { name as appName } from './app.json';
+import App from './src/App';
 
 AppRegistry.registerComponent(appName, () => App);
