+++ example/src/StyleSheet.js
@@ -1,0 +1,29 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#f5f5f5',
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginBottom: 20,
+    color: 'black',
+  },
+  box: {
+    width: 200,
+    height: 60,
+    backgroundColor: '#6200ee',
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderRadius: 10,
+    marginVertical: 10,
+  },
+  text: {
+    color: 'white', // Changed to white for better contrast
+    fontSize: 18,
+  },
+});
